requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
dnspython==2.4.2
frozenlist==1.4.0
idna==3.4
loguru==0.7.2
multidict==6.0.4
numpy==1.26.0
openai==0.28.1
pinecone-client==2.2.4
pinecone-io==0.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.6
yarl==1.9.2
